LFC193
The course helped me to understand the business better and how to address it generally. It covered all the types of open source licenses and when to use one or another, in addition to the main Linux command scripts and a good understanding of Bash. I learned a lot of good practices around working with other developers and with distributed repositories, which helped me to gain an understanding of how to collaborate with the open source community. This includes working with Git – which I do day to day now – which was necessary and important.

The most useful thing was definitely learning about Linux development tools, where I learned a lot about using command lines to manipulate files and processes using regular expressions to make tasks easier. Also, how to use various methods for debugging, compiling and building packages was very useful. Finally, understanding all the most important aspects of open source and how to collaborate with other teams is knowledge I am benefitting from.