Microsoft Innovative Educator (MIE) Expert Program
Dear Abhishek Shrivastava,

Congratulations on being selected as a Microsoft Innovative Educator Expert for 2022-2023! Each year, Microsoft selects Innovative Educator Experts to be part of this exclusive global community paving the way for their peers to share ideas, try new approaches, and learn from each other. We are inspired by your achievements and look forward to working with you throughout the one-year engagement, and beyond.

You will be joining a group of some of the most amazing educators who are empowering the students of today to create the world of tomorrow. At Microsoft, we believe that technology alone cannot build 21st century skills for students. It is an accelerator, but the power of change lies within educators. That is why we bring changemakers, like yourself, together to recognize and grow your achievements.

You have been chosen as an MIE Expert because you are self-driven, passionate about your work, have true collaborative spirit, and strive to inspire students with outside-the-box thinking on technology in education. We appreciate your resourcefulness and entrepreneurial spirit. As agents of change, you have achieved excellence in education, using technology and social media. We are grateful for the opportunity to tap into your enthusiasm and enable others to benefit from what is clearly your passion.

We understand that there is a lot of excitement and honor around being a MIE Expert; please share the enthusiasm with your school leaders, colleagues, and among friends and followers on your social media channels. We would love for you to share the great news on social media.

Use #MicrosoftEdu #MIEExpert
Example post: "This year will be (insert adjective). So excited to be selected as an MIE Expert for 2022-2023 #MicrosoftEdu #MIEExpert."
You can use the graphics found here if you want.
Actions to take immediately:

Click here and bookmark or download to OneNote to the MIE Expert Quick Start Guide (this lives in a Consumer OneDrive).
Join our private Facebook group for 2022-2023 here (use the code phrase, “I believe in equity”, and the name of your LSS: Meetu Tiwari.
Check out the products and tools (offers found in QuickStart Guide) you have access to from our partners.
Join our MIE Expert Teams site – directions are here.

Download the official signature block just for MIE Experts to use in your emails.
We will be awarding your MIE Expert badge through Credly (https://credly.com) using the email you entered for your delivery email account in the next week.
Again, CONGRATULATIONS on your selection as an MIE Expert. We are excited to be a part of your journey in creating an environment of innovation in education using technology.

Your local Microsoft team will be in contact with you shortly.

Christina Thoresen, Director, Education Engagement Programs
Sonja Delafosse, Senior Manager Educator Programs